[#update] improve implementation

+++ src/evaluationSite-Services/Evaluation.API/Interfaces/IEvaluationArticle.cs
@@ -2,8 +2,8 @@
 
 public interface IEvaluationArticle
 {
-    Task<long> CountArticlesAsync();
-    Task<long> CountArticlesByTypeAsync(int categoryId);
+    Task<int> CountArticlesAsync();
+    Task<int> CountArticlesByTypeAsync(int categoryId);
     Task<List<EvaluationArticle>> GetArticlesAsync(int pageSize, int pageIndex, string ids = null);
     Task<List<EvaluationArticle>> GetArticlesAsync(int pageSize, int pageIndex, int categoryTypeId);
     Task<EvaluationArticle> GetArticleAsync(int id);
